VFS-555:  Add the ability to specify an alternate HDFS configuration file to override any specific HDFS settings.  One use for this is to set a different value for the dfs.client.use.datanode.hostname property in order to access HDFS files stored in an AWS installation (from outside their firewall).  There are other possible uses too.
There is now an "configName" property in HdfsFileSystemConfigBuilder which
can be set in the file system options in order to set an alternate config
file.  This file can have any number of option settings which will add to
or override the default settings.  In other words, this file is loaded
after the default one (which resides in the HDFS .jar file).


git-svn-id: https://svn.apache.org/repos/asf/commons/proper/vfs/trunk@1655773 13f79535-47bb-0310-9956-ffa450edef68