VFS-586:  Add several additional ways to specify an alternate HDFS configuration.
Motivation:  the previous changes in this area allowed specifying the name of
one resource (in the CLASSPATH) for an alternate configuration.  Further use
with this feature showed the need for being able to specify more than one
resource (possibly), and to be able to load configurations from local files
as well as basically any way that HDFS configuration parameters can be
specified (according to the HDFS APIs).

Changes:  therefore, the API in HdfsFileSystemConfigBuilder was greatly
extended to allow:
a) Multiple resource names to be given;
b) File system paths to be specified (one or many);
c) URLs to be listed (multiple if desired) so configuration can be loaded
   from a web application;
d) Another HDFS Configuration object (one of these);
e) Any miscellaneous InputStream (only one of these).

And the configuration step in HdfsFileSystem was enhanced to check for
all these possible sources of configuration information and add the
available ones to the HDFS "Configuration" object during initialization.

One of the existing APIs was changed:  getConfigName -> getConfigNames
(with a corresponding change to the return type:  String -> String[]).

Note: I have not been able to figure out a good unit test for this yet,
but it is working quite well in our application.


git-svn-id: https://svn.apache.org/repos/asf/commons/proper/vfs/trunk@1720416 13f79535-47bb-0310-9956-ffa450edef68